Add files via upload
• We find Loging in and registering into any web page that is present in the Internet, where there are different platforms 
and every platform has it’s own users. In maximum all of the websites, the website asks the user to register themselves 
on the website or if they do have an account they ask us to login into our account. 
• The login And Registration System project in C++ involves mainly the user registration process. User Credentials like 
usernames and passwords are asked from the user. If the registration of the user is successful then with the given 
credentials a file will be created of a particular user in the database. 
• The purpose of this project is to demonstrate the fundamental concepts of file handling, user authentication, and basic 
security measures in C++. This project encompasses user registration, login functionality, and even includes a password 
reset feature.
• The code that we are building for this purpose, has 3 options displayed on the output console to either login or register 
or to reset a password by giving the username as input. There is a menu displayed in the output console
• This mini project contains limited features, but the essential one. The system does not create an external 
file to store the user’s data permanently. This system is in C++ Programming Language and different 
variables, strings have been used for the development of it. 
• This mini project demonstrates the practical use of a programming language its features as well as to 
generate an application which can be used interest to develop a good user friendly website where he can 
register himself, make himself a member and keep progressing in his website. 
• The scope of this project is in such a way that this can be used in any website that is being developed on 
the internet, who wants to keep track on their users.